Move instance related functionality into model
